UI/UX changes in RegisterActivity.kt
RegisterActivity.kt is made such as it
takes different input and works different for
buyers on sellers.

User can switch between registering as a buyer or
a seller with buttons.

The different EditTexts and Button change visibility
depending on the selected option between buyer and seller.

Functionality is yet to be added.